Redid Attachment System and Attachment Detection
The way that blocks attach to one anouther is now a complelty different system. This new system is much more stable and less prone to having blocks get flung about to the edges of the known universe.

In other news, blocks can now detect if they are connected to the core by a long string of connections (more correctly the core can tell them if they are connected). This means removing a block will cause any other blocks that were connecting through it to drop off.

Overall a very good commit.